Variable Explorer: Move dicteditor to varexp namespace
