+++ spyderlib/widgets/varexp/dicteditor.py
